task3 - use parseFloat instead of parseInt

+++ js/task-3.js
@@ -1,7 +1,7 @@
 function getElementWidth(content, padding, border) {
-  const contentSize = Number.parseInt(content);
-  const paddingSize = Number.parseInt(padding);
-  const borderSize = Number.parseInt(border);
+  const contentSize = Number.parseFloat(content);
+  const paddingSize = Number.parseFloat(padding);
+  const borderSize = Number.parseFloat(border);
 
   return contentSize + 2 * paddingSize + 2 * borderSize;
 }
